+++ python/cudf/cudf/core/column/column.py
@@ -711,8 +711,6 @@ class ColumnBase(Column):
                 f"to isin(), you passed a [{type(values).__name__}]"
             )
 
-        from cudf import DataFrame, Series
-
         lhs = self
         rhs = None
 
@@ -720,6 +718,10 @@ class ColumnBase(Column):
             # We need to convert values to same type as self,
             # hence passing dtype=self.dtype
             rhs = as_column(values, dtype=self.dtype)
+
+            # Short-circuit if rhs is all null.
+            if lhs.null_count == 0 and (rhs.null_count == len(rhs)):
+                return as_column(cupy.zeros(len(self), dtype="bool"))
         except ValueError:
             # pandas functionally returns all False when cleansing via
             # typecasting fails
@@ -729,14 +731,8 @@ class ColumnBase(Column):
         if is_categorical_dtype(lhs):
             lhs_cats = lhs.cat().categories._values
             rhs_cats = rhs.cat().categories._values
-            if np.issubdtype(rhs_cats.dtype, lhs_cats.dtype):
-                # if the categories are the same dtype, we can combine them
-                cats = Series(lhs_cats.append(rhs_cats)).drop_duplicates(
-                    ignore_index=True
-                )
-                lhs = lhs.cat().set_categories(cats, is_unique=True)
-                rhs = rhs.cat().set_categories(cats, is_unique=True)
-            else:
+
+            if not np.issubdtype(rhs_cats.dtype, lhs_cats.dtype):
                 # If they're not the same dtype, short-circuit if the values
                 # list doesn't have any nulls. If it does have nulls, make
                 # the values list a Categorical with a single null
@@ -745,8 +741,8 @@ class ColumnBase(Column):
                 rhs = as_column(pd.Categorical.from_codes([-1], categories=[]))
                 rhs = rhs.cat().set_categories(lhs_cats).astype(self.dtype)
 
-        lhs = DataFrame({"x": lhs, "orig_order": cupy.arange(len(lhs))})
-        rhs = DataFrame({"x": rhs, "bool": cupy.ones(len(rhs), "bool")})
+        lhs = cudf.DataFrame({"x": lhs, "orig_order": cupy.arange(len(lhs))})
+        rhs = cudf.DataFrame({"x": rhs, "bool": cupy.ones(len(rhs), "bool")})
         res = lhs.merge(rhs, on="x", how="left").sort_values(by="orig_order")
         res = res.drop_duplicates(subset="orig_order", ignore_index=True)
         res = res._data["bool"].fillna(False)
@@ -868,7 +864,7 @@ class ColumnBase(Column):
             codes=labels._column,
             mask=self.mask,
             ordered=ordered,
-        )
+        ).astype(dtype)
 
     def as_numerical_column(self, dtype, **kwargs):
         raise NotImplementedError
@@ -1129,8 +1125,6 @@ def build_categorical_column(
     ordered : bool
         Indicates whether the categories are ordered
     """
-    if len(categories) == 0 and len(codes):
-        raise ValueError("Cannot have nonempty codes for empty categories")
 
     dtype = CategoricalDtype(categories=as_column(categories), ordered=ordered)
 
@@ -1246,7 +1240,9 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
         if dtype is not None:
             col = col.astype(dtype)
 
-        if np.issubdtype(col.dtype, np.floating):
+        if isinstance(col, cudf.core.column.CategoricalColumn):
+            return col
+        elif np.issubdtype(col.dtype, np.floating):
             if nan_as_null or (mask is None and nan_as_null is None):
                 mask = libcudf.transform.nans_to_nulls(col.fillna(np.nan))
                 col = col.set_mask(mask)
@@ -1299,6 +1295,7 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
                 categories = as_column([], dtype="object")
             else:
                 categories = as_column(arbitrary.dictionary)
+
             dtype = CategoricalDtype(
                 categories=categories, ordered=arbitrary.type.ordered
             )
@@ -1401,12 +1398,11 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
         elif arbitrary.dtype == np.bool:
             data = as_column(cupy.asarray(arbitrary), dtype=arbitrary.dtype,)
         elif arbitrary.dtype.kind in ("f"):
-            arb_dtype = check_cast_unsupported_dtype(
-                arbitrary.dtype if dtype is None else dtype
-            )
+            arb_dtype = check_cast_unsupported_dtype(arbitrary.dtype)
             data = as_column(
                 cupy.asarray(arbitrary, dtype=arb_dtype),
                 nan_as_null=nan_as_null,
+                dtype=dtype,
             )
         elif arbitrary.dtype.kind in ("u", "i"):
             data = as_column(
@@ -1560,7 +1556,7 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
             except (pa.ArrowInvalid, pa.ArrowTypeError, TypeError):
                 if is_categorical_dtype(dtype):
                     sr = pd.Series(arbitrary, dtype="category")
-                    data = as_column(sr, nan_as_null=nan_as_null)
+                    data = as_column(sr, nan_as_null=nan_as_null, dtype=dtype)
                 elif np_type == np.str_:
                     sr = pd.Series(arbitrary, dtype="str")
                     data = as_column(sr, nan_as_null=nan_as_null)