+++ python/dask_cudf/dask_cudf/backends.py
@@ -3,8 +3,10 @@ import numpy as np
 import pandas as pd
 import pyarrow as pa
 
+from dask.dataframe.categorical import categorical_dtype_dispatch
 from dask.dataframe.core import get_parallel_type, make_meta, meta_nonempty
 from dask.dataframe.methods import concat_dispatch
+from dask.dataframe.utils import UNKNOWN_CATEGORIES
 
 import cudf
 from cudf.utils.dtypes import is_string_dtype
@@ -56,7 +58,9 @@ def _nonempty_index(idx):
 def _get_non_empty_data(s):
     if isinstance(s._column, cudf.core.column.CategoricalColumn):
         categories = (
-            s._column.categories if len(s._column.categories) else ["a"]
+            s._column.categories
+            if len(s._column.categories)
+            else [UNKNOWN_CATEGORIES]
         )
         codes = cp.zeros(2, dtype="int32")
         ordered = s._column.ordered
@@ -121,6 +125,11 @@ def concat_cudf(
 ):
     assert join == "outer"
     return cudf.concat(dfs, axis=axis, ignore_index=ignore_index)
+
+
+@categorical_dtype_dispatch.register((cudf.DataFrame, cudf.Series, cudf.Index))
+def categorical_dtype_cudf(categories=None, ordered=None):
+    return cudf.CategoricalDtype(categories=categories, ordered=ordered)
 
 
 try: